adding test slashing
