+++ x/slashing/init_test.go
@@ -1,0 +1,9 @@
+package slashing_test
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+var (
+	InitTokens = sdk.TokensFromConsensusPower(200)
+)
